VHM: Use refcnt and idr mechanism to manage ioreq_client
Currently the ioreq_client is managed by using static array. it is possible that the
ioreq_client is still accessed while it is released in another thread. In such case
it will have the race-condition.

In order to avoid the race-condition, the refcnt is used to indicate whether it can
be released. And the IDR(alloc/find/replace) is used to manage the allocation/lookup/
destroy of ioreq_client(spin_lock_bh is used).

MAX_CLIENT is increased from 64 to 1024 after ioreq_client is managed by IDR.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>